move tests to test_view
